Added Search To Member

+++ app.js
@@ -17,12 +17,12 @@ const app = express();
 // ========= server =========
 mongoose.set("strictQuery", true);
 mongoose
-.connect(process.env.MongoUrl)
-.then(() => {
-  console.log("DB connected");
-  app.listen(process.env.PORT || 8080, () => console.log(`listening on http://localhost:${process.env.PORT}`));
-})
-.catch((error) => console.log(`DB connection error ${error}`));
+  .connect(process.env.MongoUrl)
+  .then(() => {
+    console.log("DB connected");
+    app.listen(process.env.PORT || 8080, () => console.log(`listening on http://localhost:${process.env.PORT}`));
+  })
+  .catch((error) => console.log(`DB connection error ${error}`));
 
 //============server=========
 app.use(cors());
@@ -41,20 +41,17 @@ app.use(memberRouter);
 app.use(bookRouter); //  /books and /categories routes
 
 // not found Middleware
-app.use((req,res,next)=>{
-    res.status(404)
-    .json({massage:"page not found"})
-    if(req.file && req.file.path)
-        fs.unlinkSync(req.file.path);
-})
+app.use((req, res, next) => {
+  res.status(404).json({ massage: "page not found" });
+  if (req.file && req.file.path) fs.unlinkSync(req.file.path);
+});
 
 // error Middleware
-app.use((err,req,res,next)=>{
-    let status=err.status|| 500;
-    res.status(status).json({message:err+""});
-    // Delete Saved files in case of error throwing 
-    if(req.file && req.file.path)
-        fs.unlinkSync(req.file.path);
-})
+app.use((err, req, res, next) => {
+  let status = err.status || 500;
+  res.status(status).json({ message: err + "" });
+  // Delete Saved files in case of error throwing
+  if (req.file && req.file.path) fs.unlinkSync(req.file.path);
+});
 
 module.exports = app;
